* commit '464a8d96ba1723517254d5497b0301b283a2e282':
  fix: fix the quantity filed type to Long
  fix: fix the quantity filed type to Long

+++ src/main/java/app/entity/Cart.java
@@ -14,11 +14,11 @@ public class Cart {
   @NonNull
   private Long productId;
   @NonNull
-  private Integer productQuantity;
+  private Long productQuantity;
 
 
   public static Cart CartCompKeyBuilder(ProductAndMemberCompositeKey productAndMemberCompositeKey,
-      Integer productQuantity) {
+      Long productQuantity) {
     return new Cart(productAndMemberCompositeKey.getMemberId(),
         productAndMemberCompositeKey.getProductId(), productQuantity);
